Listen for hashchange events.
Now, when the location.hash changes, the map center and zoom will change
accordingly. This is particularly nice for pasting in a URL that only differs by
hash. Also, the hash control now ignores the hash if any part of it isNaN, or if
it contains fewer than three components.