New layer `visible` property.
A layer's visibility can now be toggled without adding or removing the layer
from the map. This avoids inadvertantly changing the z-order of layers. In
addition to removing the tiles from the SVG image, the layer container's
visibility attribute is set to "hidden", such that if the user has extended the
layer with additional elements, they become hidden as well.