Rotation!
Maps can now be rotated using the `angle` property. For simplicity, the layer
implementation currently computes the bounding box to fill the rotated viewport;
in a future commit we will use a scanline fill algorithm to avoid loading tiles
outside the viewport.

This commit also fixes a bug in the `coordinatePoint` and `pointCoordinate`
methods, which were treating points pre-fractional zoom. Now points are always
considered to be in screen space (pixel coordinates), consistent with the
`locationPoint` and `pointLocation` methods. Note, however, that the
tile-specific projection used by layers is still pre-transform.