Improve stability of mousewheel interaction.
Previously, the location under the mouse could drift as the mousewheel was used
to zoom in and out. Now we compute the location under the mouse on mousemove, so
that it does not drift when the map is zoomed. The location is specified as an
optional third argument to `map.zoomBy`.

The `map.panBy` method is reimplemented to be slightly faster and more accurate
by avoiding an extra point <-> location conversion.

Most importantly, the hash control now ignores the 'hashchange' event that it
generates when it changes the hash. Previously, this was causing noticeable
aliasing on smooth zoom as any movement of the map would be rounded to the
precision of the hash control.