Improve stability of mousewheel interaction (again).
The previous fix suffered from two problems.

First, redundant hashchange events are occasionally dispatched by the browser,
and later events would not be ignored; thus the hash control would end up moving
the map on map move. We avoid this now by caching the location hash, and only
moving the map when the hash changes. This also allows us to ignore the
hashchange event when the map moves.

Second, the wheel control was caching the location under the mouse, and only
updating it when the mouse moves. However, the map can move independently of the
mouse, for example with arrow controls or animation. Thus, we now clear the
cached location when the map moves. (And, we're careful to preserve the cached
location when we zoom the map.)