Use scanline algorithm to determine visible tiles.
Previously, if the map were rotated, we would compute the axis-aligned bounding
box to determine which tiles were visible. This is easy, but has the undesirable
property that it can load (and display) many tiles that are inside the bounding
box but outside the viewport.

Now we use the triangle scanline algorithm to determine exactly which tiles are
visible. The algorithm is modified slightly so that it displays any pixel (i.e.,
tile) that intersects the triangle.