Fix two bugs with proxy tiles.
With zoom transforms, it was possible for a proxy tile to be less than -4 or
greater than +2 zoom levels from the map zoom level. This caused Polymaps to
crash because the corresponding layer `g` element was missing. Now we properly
shift the proxy zoom levels when a zoom transform is used.

Another problem with proxy tiles is that we were accidentally doubling the proxy
count for tiles along the viewport diagonal, per the new scanline algorithm. We
now compute the correct proxy count, which is necessary so that the proxy tiles
are removed when the referenced tiles load.