Change the semantics of `map.centerRange`.
Previously, the centerRange constraint would directly restrict the allowed
latitudes and longitude of the map center. This was not particularly useful,
since usually you want to restrict the viewable area of the map, rather than the
center point. In addition, the map had an internal constraint that the visible
latitudes be between [-85.05112877980659, 85.05112877980659]. If zoomed out such
that this entire range were visible, the map would be centered vertically. There
was also a bug if the map were rotated at certain angles where this constraint
would not be enforced correctly.

The new meaning of centerRange is the "viewable" area of the world, and defaults
to [

  {lat: -85.05112877980659, lon: -Infinity},
  {lat: 85.05112877980659, lon: Infinity}

], which subsumes the previous internal constraint. Thus, the internal
constraint can now be disabled by setting the centerRange to null. If a
centerRange is set, then at least some of the area within the centerRange is
guaranteed to be visible. Note, however, that this visible area may be
infinitely small. Thus, if the user pushes the map off the screen entirely, it
is always immediately off the edge of the viewport, easily available to be
brought back into view.

The motivation for this commit is three-fold: First, it allows the internal
constraint to be disabled, which is particularly useful for cases where the map
tiles aren't really spherical mercator projections of the Earth. Second, it
allows panning restrictions to be defined in a way that is more convenient than
under previous semantics. Lastly, it guarantees that when zooming in (say, via
mousewheel or dblclick), the visible location under the mouse is guaranteed to
stay under the mouse after the zoom level changes.