Better mousewheel speed heuristic.
Unfortunately, mousewheel speeds are horribly broken. See this thread

  https://lists.webkit.org/pipermail/webkit-dev/2010-June/013152.html

The new heuristic is designed primarily for discrete devices which return
wheelDeltas as multiples of 120. In WebKit nightly continuous devices return
wheelDeltas consistent with discrete devices.

The problem is that in Safari 5 and Chrome the wheelDeltas on continuous devices
are pre-multiplied by 120, returning ridiculously large numbers. So for these
browsers we now keep a moving average of the apparent mousewheel speed, and if
above some threshold, assume a continuous device and dampen accordingly.