Don't include HEAD revision in polymaps[.min].js.
When these generated files were not part of the repository, it made sense to
include the HEAD revision id, so that we knew how the files were built. These
IDs are no longer needed as Semantic Versioning gives us the same information
(and more!).

Furthermore, now that the generated files are part of the repo, the HEAD
revision would be the commit prior to the one that included the generated files,
which is both confusing and causes spurious modifications if the files are
rebuilt after pulling.

Note: since this is a minor comment-only change, I'm not bumping the version.