Fix two bugs in Hash and Queue, respectively.
The Hash control was attempting to use precision -Infinity for zoom level zero.
Now it uses precision 0. The Queue's image method was not passing the loaded
image to the callback function, unlike the other text, xml and json methods.
Having the image available to the callback allows the implementation of post-
processing with canvas, for example.