Better browser workarounds.
Rollback commit 73168a7, which didn't entirely solve the problem for Firefox and
introduced new artifacts in WebKit browsers. The root problem is that Firefox
positions the SVG image with subpixel accuracy (as evinced by getScreenCTM()
returning subpixel values for the `e` and `f` attributes), and so even if the
SVG internally positions images at even pixel boundaries, Firefox still has
anti-aliasing artifacts. A proper solution to this problem is trickier than I
thought, and since it's only a minor aesthetic problem I'm going to ignore it
for now.

The `wheel` control now triggers the wheel speed dampening if a single large
wheelDelta is seen, rather than trying to compute a moving average of the wheel
speed. This means that false positives are more annoying, but false negatives
are significantly less likely, so this should be a significant improvement for
Safari and Chrome.

Lastly, there's a workaround for the broken getScreenCTM in WebKit browsers when
the window is scrolled! Hooray.