Fix two bugs in tile visibility determination.
First, if the zoom level (after transformation) is less than 0, set the minimum
and maximum visible row to 0, rather than 1 << -1 (-2147483648). Second, fix a
bug in the triangle scanline algorithm where we were not properly determining
which of the two spans was the "left" span. We now determine at which point the
spans touch, and compare the x-coordinate at the opposite y-value.