Use keypress events for arrow zooming.
The plus and minus key codes are not consistent across browsers. The char codes
appear to be more consistent, so we use those; note that this also has the
side-effect of enabling repeated zooming if the plus or minus key is held down.
This makes zooming more consistent with the current (custom repeating) pan
behavior.