Shift-drag to set map extent.
This functionality is added to the `compass` control, which at first glance
might seem weird, but it provides a convenient place for the dragging rect (on
top of the other layers), and can also inherit the compass styling. This
functionality is built on top of the map's new `extent` method, which takes a
two-element location array, similar to `centerRange`.

In a future change, it would be nice to add key listeners to the compass
control, so that the escape key can cancel the shift-drag operation, and the
space key can move the bounds rather than resize them. You might also want to
floor the zoom level after setting the extent. And maybe a way to turn off
shift-dragging if the functionality is not desired.