Allow geoJson.clip to be set after add.
Setting geoJson.tile(false) also explicitly, rather than implicitly, disables
clipping. The size of the clipping rect is inferred from the tile size when the
map is set (as before); this may change in the future if the map dispatches an
event when the tile size changes.

Also in this commit, the features associated with a geoJson tile are now
accessible as the `features` attribute of the tile. This is distinct from the
`features` attribute of the event, which only includes the updated features.