Use getBBox instead of getClientRects.
After some browser testing and a closer reading of the CSSOM specification, it
turns out Firefox was correctly reporting the size of the tight SVG bounding
box, while WebKit was returning (incorrectly) the size of the SVG container. We
now use `rect.getBBox` on all browsers, which is much more reliable!