Fix a bug with default geoJson settings.
The default URL for geoJson layers was "about:blank"; this causes a security
exception when issuing an XMLHttpRequest from another domain. The new default
URL is null, which avoids any XMLHttpRequest and just creates blank tiles. In
addition, setting the features property to a non-null value has the side-effect
of setting the URL to null and vice versa.

For consistency, the default URL for image layers is now also null, which causes
no image to be loaded. This commit also fixes a bug when the URL is a string
with no placeholders, allowing simple repeating image tiles to be created.