Port polymaps.org to gh-pages.
Note that the previous version still lives in the www folder of the master
branch; that will be removed in a future commit. Also note that the examples
will continue to live in the master branch; the code is copied to the gh-pages
branch as part of the build process.