Use .PHONY targets.
Ideally, I'd like to be able to track dependencies between files in the master
branch and copies of those files (and derived versions) in the gh-pages
branch. However, it's a bit tricky to do so because using git checkout to
switch branches changes the modification date of files. So instead the
Makefile now uses special phony targets, so that all the files are copied
whenever make is run. Since make is only run to copy the files from master to
gh-pages, this seems like an acceptable limitation.