A different hack for reliable mousewheel.
Wheel events are utterly broken. Even though WebKit's mousewheel precision bug
has been fixed, Chrome and Safari still differ in terms of mousewheel
acceleration. Safari and Firefox report the correct value (the number of pixels
that would be scrolled), but Chrome does not. I've not tested Opera or IE9.

Fortunately, we can determine exactly the number of pixels that would be
scrolled by dispatching the mousewheel event to a scrollable area, and then
observing the change in scrollTop. However, dispatching the received event is
not strictly allowed, so ignore this error and hope for the best.